Fix: remove duplicate common exam

+++ src/exam/model/exam-score.model.ts
@@ -1,11 +1,11 @@
 import { BaseModel } from '@common/model/base.model';
-import { Field, Int, ObjectType } from '@nestjs/graphql';
+import { Field, Float, Int, ObjectType } from '@nestjs/graphql';
 
 @ObjectType()
 export class ExamScore extends BaseModel {
   @Field(() => Int)
   problemNumber: number;
 
-  @Field(() => Int)
+  @Field(() => Float)
   maxScore: number;
 }
